+++ src/functions/events/filterEvents.js
