+++ src/functions/events/getSubscriptionStatus.js
