+++ src/functions/events/index.js
@@ -1,0 +1,11 @@
+import { filterEvents, selectUserEvents, sortEvents } from './filterEvents';
+import { validateEvent } from './validateEvent';
+import { getSubscriptionStatus } from './getSubscriptionStatus';
+
+export {
+  filterEvents,
+  selectUserEvents,
+  sortEvents,
+  validateEvent,
+  getSubscriptionStatus
+};
