+++ src/functions/events/validateEvent.js
